Move createTrustedClusterPair instead
